Remodel TDrawSurface and TSurfaceView
Preserving the surface's contents upon resizing serves no useful use case, since they will look distorted. So we can remove the fillColor field and even avoid using realloc when increasing the surface size.

TSurfaceView does need a fill color which cannot be provided as argument to draw(). So make it a member variable of this class.