platform: replace FdInputStrategy with EventWaiter/EventSource
The new EventWaiter does not rely on global variables and is cross-platform. EventSources can be wrapped (decorated) in a cleaner way than FdInputStrategy (see LinuxConsoleInput). This brings some class design improvements. For example, SigwinchHandler can now be managed by the PlatformStrategy directly instead of requiring to be inherited by an InputStrategy.