BufferedDisplay: optimize out redundant screen buffer
We need two screen buffers: one with the current screen state (BufferedDisplay::buffer) and one with the screen state in the last flush (BufferedDisplay::flushBuffer). The public API requires a buffer to be located at TScreen::screenBuffer and until now we allocated a third buffer for that variable. But it is redundant and can be avoided in order to consume less memory and skip a memcpy operation in BufferedDisplay::screenWrite.

From now on, Platform::reloadScreenInfo should only be invoked when setting TScreen::screenBuffer, given that references to BufferedDisplay::buffer may be invalidated.