platform: auto-detect terminal size changes in event loop
Before 10e16950ca3f0969d5e37dacec6cb4fbcbd33d98, Platform::reloadScreenInfo used to be invoked from Platform::setUpConsole so that Platform would report up-to-date terminal dimensions. That is no longer the case and terminal dimensions in BufferedDisplay may be out-of-date in certain cases (for example, after recovering from SIGTSTP). To solve this we check for size changes every time when reading events (and not just in signal handlers or when receiving WINDOW_BUFFER_SIZE_EVENT).