Windows: bind StdioCtl lifetime to ConsoleStrategy instead of Platform
On Windows we must only allocate a console when the screen is active. On Unix this is not only unnecessary but also impossible since we need to check for a Linux console in Platform's constructor. So we can leave a StdioCtl member in Platform on Unix. For the console size we introduce a DisplayStrategy method to avoid having to check whether we have a StdioCtl object on hand or not.