Improve performance of TCellChar
We avoid creating temporary TCellChar objects by using member methods instead of constructors. Although the result is the same, this helps the compilers optimize TCellChar usages.

The 3 bytes that were previously unused in TScreenCell are now used for storing text in TCellChar. So besides the optimizations, we can now fit more combining characters in a TCellChar.