Replace pointer casting in __DELTA macro with dynamic_cast<void *> in ipstream
The __DELTA macro was an attempt to compute at compile-time something which in some cases is only known at runtime. This would result in a segmentation fault when attempting to dereference an invalid pointer. One way to do the same without undefined behaviour or changing the API is dynamic_cast<void *>, which dynamically computes the address of the most derived class.

The original Turbo Vision couldn't use dynamic_cast because it was not introduced until Borland C++ 4.0. Since we support Borland C++ 4.52, we are free to use it.