platform: filter out unrecognized bits in dwControlKeyState
putty4far2l (and possibly other applications) use more bits than the ones defined in the Win32 API. See https://github.com/microsoft/terminal/issues/337.

This conflicts with our flags.

In MS-DOS this is not a problem because THardwareInfo::shiftState() returns a char and all the bits are already in use. Therefore, we can safely reassign kbPaste to the value 0x100.

Closes #92.
Fixes magiblot/turbo#47.