StderrRedirector: use pipe buffer instead of temporary file
Temporary files may lead to unbound memory or disk usage. This can make the whole system unusable.

Conversely, a pipe buffer has a limited size. Write operations will fail once it is full; this may cause the process to hang if it tries to write into the pipe relentlessly (which is not the case of stdio functions), but even then it is much better than crashing the whole system.