platform: find a universal solution for consuming unprocessed input
The problem that we fixed for far2l (#49) can be also reproduced with Conpty and win32-input-mode.

I figured out that we can get any terminal to reply to us with a Device Status Report. So use that instead of relying on the far2l protocol.

This commit also fixes a bug when consuming unprocessed input. With a slow connection, we would fail to read the terminal response, which is why we were forced to add a timeout in 08d83577d47f65b8c9d5d05d965deb206ff39648. This was caused by calling NcursesInput::getEvent, which would at times interpret the response sequence as an Alt+Key combination. In order to prevent confusion when debugging, wgetch is now only called directly from NcursesInputGetter.