platform: do not destroy the Platform instance on exit
Otherwise, secondary threads calling TEventQueue::wakeUp() could cause the application to crash on exit.

We do it this way because there is no safe way to destroy the Platform orderly. In theory it could be done with thread_local shared pointers, but this appears to be broken in some compilers (e.g. MINGW).