platform: reorganize character processing functions
By using a struct, adding more functions is now simpler. Since Platform's destructor was removed, these functions no longer need to be mixed with any other Platform functions in a single translation unit.

Previously there was a static 'charWidth' function in every console adapter that was defined separately from the other class functions, which was weird.

The override of the global CharOps in the tests has also been reorganized, so that it can be reused from other tests.