Polish TEditor's API extensions
While implementing new behaviour, additional public functions and fields were added to TEditor. Some of those functions were not well conceived in the sense that their behaviour was not clear and they only added complexity or inconsistency to the API.

This commit minimizes the amount of new public functions in TEditor. For example, instead of having an 'insertMultilineText' method which converts line endings on-the-fly, this behaviour has been integrated into the original 'insertBuffer' function.